Override console.error in Failure test.
The failure test throws an exception on purpose, so we can
catch it and render an ad-hoc component.
However, jest would still output the exception when running the
test, which can be a bit confusing since the test is still passing.
Overriding console.error for the time of the test allow us to
get rid of the exception message.